[GR-47080] Add support for TimeStampNode intrinsic.
